+++ src/routes/benefitRoutes.js
@@ -1,6 +1,5 @@
 const express = require('express');
 const {
-  createBenefit,
   getBenefitList,
   updateBenefit,
 } = require('../controllers/benefitControllers');
@@ -9,7 +8,6 @@ const formatApiResponse = require('../middlewares/formatApiResponse');
 
 const router = express.Router();
 
-router.post('/create', authenticateToken, formatApiResponse, createBenefit);
 router.get('/', authenticateToken, formatApiResponse, getBenefitList);
 router.put('/status', authenticateToken, formatApiResponse, updateBenefit);
